Backend/requirements.txt
flask
flask-cors
python-dotenv
cohere
